requirements.txt
adblockparser==0.7
aiomultiprocess==0.8.0
attrs==19.3.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.1
blinker==1.4
Brotli==1.0.9
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
fake-useragent==0.1.11
Flask==1.1.2
fonttools==4.14.0
funcparserlib==0.3.6
gevent==20.6.2
greenlet==0.4.16
h11==0.10.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
kaitaistruct==0.8
kiwisolver==1.2.0
ldap3==2.7
lxml==4.5.1
maoyan-font==2018.12.1
MarkupSafe==1.1.1
matplotlib==3.3.1
mitmproxy==5.2
nltk==3.5
numpy==1.19.0
parsel==1.6.0
parso==0.7.1
passlib==1.7.2
pickleshare==0.7.5
Pillow==7.2.0
priority==1.3.0
prompt-toolkit==3.0.7
Protego==0.1.16
protobuf==3.11.3
psutil==5.7.2
publicsuffix2==2.20191221
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pydivert==2.1.0
PyExecJS==1.5.1
Pygments==2.7.1
PyHamcrest==2.0.2
pymaoyanfont==0.0.1
pymongo==3.11.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pyquery==1.4.1
pytesseract==0.3.5
python-dateutil==2.8.1
qt5reactor==0.6.1
queuelib==1.5.0
redis==3.5.3
regex==2020.6.8
requests==2.23.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.5.2
Scrapy==2.1.0
scrapy-redis==0.6.8
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
sortedcontainers==2.1.0
soupsieve==2.0.1
splash==3.5
tesserocr==2.4.0
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
Twisted==20.3.0
urllib3==1.25.9
urwid==2.0.1
w3lib==1.22.0
wcwidth==0.2.5
wechatarticles==0.4.8
Werkzeug==1.0.1
wsproto==0.15.0
wxPython==4.1.0
XlsxWriter==1.2.9
xvfbwrapper==0.2.9
zope.event==4.4
zope.interface==5.1.0
zstandard==0.13.0
